port r68743
git-svn-id: https://svn.r-project.org/R/branches/R-3-2-branch@68800 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/connections.c
@@ -4039,7 +4039,7 @@ SEXP attribute_hidden do_readbin(SEXP call, SEXP op, SEXP args, SEXP env)
 			break;
 #if HAVE_LONG_DOUBLE && (SIZEOF_LONG_DOUBLE > SIZEOF_DOUBLE)
 		    case sizeof(long double):
-			REAL(ans)[i] = u.ld;
+			REAL(ans)[i] = (double) u.ld;
 			break;
 #endif
 		    default:
